requirements.txt
argparse (1.2.1)
coverage (3.7.1)
Django (1.6.5)
django-braces (1.4.0)
django-debug-toolbar (1.2.1)
django-piston (0.2.3)
django-redis (3.6.2)
django-wysiwyg (0.5.1)
easy-thumbnails (2.0.1)
gunicorn (18.0)
ipython (2.1.0)
itsdangerous (0.24)
MarkupSafe (0.23)
pilkit (1.1.9)
Pillow (2.4.0)
pip (1.5.4)
psycopg2 (2.5.3)
python-ptrace (0.7)
redis (2.10.1)
requests (2.3.0)
selenium (2.42.1)
setuptools (2.2)
six (1.6.1)
South (0.8.4)
sqlparse (0.1.11)
